\name{design.MSPRT_oneZ}
\alias{design.MSPRT_oneZ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: Designing the MSPRT for one-sample z tests
}
\description{
\code{\link{design.MSPRT}} calls this function for designing the MSPRT in one-sample z tests. Users please refer to \code{\link{design.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
