\name{design.MSPRT_twoT}
\alias{design.MSPRT_twoT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: Designing the MSPRT for two-sample t tests
}
\description{
\code{\link{design.MSPRT}} calls this function for designing the MSPRT in two-sample t tests. Users please refer to \code{\link{design.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}

