\name{implement.MSPRT_oneZ}
\alias{implement.MSPRT_oneZ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: Implementing the MSPRT in one-sample z tests
}
\description{
\code{\link{implement.MSPRT}} calls this function for implementing the MSPRT in one-sample z tests. Users please refer to \code{\link{implement.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
