\name{implement.MSPRT_twoT}
\alias{implement.MSPRT_twoT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: Implementing the MSPRT in two-sample t tests
}
\description{
\code{\link{implement.MSPRT}} calls this function for implementing the MSPRT in two-sample t tests. Users please refer to \code{\link{implement.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
