\name{OCandASN.MSPRT_oneProp}
\alias{OCandASN.MSPRT_oneProp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: OC and ASN of a designed MSPRT in one-sample proportion tests
}
\description{
\code{\link{OCandASN.MSPRT}} calls this function for obtaining the OC and ASN of a designed MSPRT for one-sample proportion tests. Users please refer to \code{\link{OCandASN.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
