\name{OCandASN.MSPRT_oneT}
\alias{OCandASN.MSPRT_oneT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: OC and ASN of a designed MSPRT in one-sample t tests
}
\description{
\code{\link{OCandASN.MSPRT}} calls this function for obtaining the OC and ASN of a designed MSPRT for one-sample t tests. Users please refer to \code{\link{OCandASN.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
