\name{OCandASN.MSPRT_twoT}
\alias{OCandASN.MSPRT_twoT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: OC and ASN of a designed MSPRT in two-sample t tests
}
\description{
\code{\link{OCandASN.MSPRT}} calls this function for obtaining the OC and ASN of a designed MSPRT for two-sample t tests. Users please refer to \code{\link{OCandASN.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
