\name{OCandASN.MSPRT_twoZ}
\alias{OCandASN.MSPRT_twoZ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal MSPRT function: OC and ASN of a designed MSPRT in two-sample z tests
}
\description{
\code{\link{OCandASN.MSPRT}} calls this function for obtaining the OC and ASN of a designed MSPRT for two-sample z tests. Users please refer to \code{\link{OCandASN.MSPRT}}.
}
\author{
Sandipan Pramanik, Valen E. Johnson and Anirban Bhattacharya
}
