\name{MSPRT-package}
\alias{MSPRT-package}
\alias{MSPRT}
\docType{package}
\title{
\packageTitle{MSPRT}
}
\description{
\packageDescription{MSPRT}
}
\details{

  \tabular{ll}{
    Package: \tab MSPRT\cr
    Type: \tab Package\cr
    Version: \tab 3.0\cr
    Date: \tab 11-11-2020\cr
    License: \tab GPL>=2\cr
  }
}
\author{
\packageAuthor{MSPRT}

Maintainer: \packageMaintainer{MSPRT}
}